+++ src/components/form/SubmitButton.tsx
@@ -1,10 +1,16 @@
-import S from "./styles/form.module.css";
+import S from "../styles/form.module.css";
+
 interface Props {
   label: string;
   type?: "button" | "submit" | "reset";
   disabled?: boolean;
 }
+
 function SubmitButton({ label, type = "button", disabled = false }: Props) {
-  return <button type={type} className={S.submitButton} disabled={disabled}>{label}</button>;
+  return (
+    <button type={type} className={S.submitButton} disabled={disabled}>
+      {label}
+    </button>
+  );
 }
 export default SubmitButton;
